Update LoginScreen.java
I have a working action listener for each number key, but i have it as 10 individual listeners.  I'm going to add it to your code here but commented out.  If you can find a way to get the listener for the numbers to work in a loop or something I'm listening.